sporty_panadas/requirements.txt
openai~=0.27.8
python-dotenv~=1.0.0
transformers==4.31.0
datasets~=2.13.1
peft==0.4.0
accelerate==0.21.0
bitsandbytes==0.40.2
safetensors>=0.3.1
tokenizers>=0.13.3
scipy==1.11.1
wandb
